feat(categories page)
