+++ lib/core/config/routes/route_names.dart
@@ -1,12 +1,17 @@
 class RouteNames {
+  // Onboarding
   static const String onboarding = '/onboarding';
   static const String splashScreen = '/splash';
   static const String login = '/login';
   static const String signUp = '/signup';
 
+  // Navigation Bar
   static const String homePage = '/homepage';
   static const String favouritesPage = '/favouritespage';
   static const String cartPage = '/cartpage';
   static const String categoriesPage = '/categoriespage';
   static const String profilePage = '/profilepage';
+
+  // Other Pages
+  static const String fruitsCategoryPage = '/fruitscategorypage';
 }
