+++ lib/features/categories/presentation/models/category_page_model.dart
@@ -1,0 +1,8 @@
+// Mock Riverpod StateNotifier and Provider for Categories
+class Category {
+  final String title;
+  final int itemCount;
+  final String imagePath;
+  Category(
+      {required this.title, required this.itemCount, required this.imagePath});
+}
